refactor(connectors): split OAuth strategies by provider
Move the provider strategy registry into per-provider modules so Linear, Notion and custom OAuth behavior stays vertically organized. Extract shared metadata helpers to break the strategy/oauth circular dependency without changing runtime behavior.
