+++ apps/web/src/lib/connectors/oauth.ts
@@ -1,8 +1,8 @@
 import crypto from 'node:crypto'
 
+import { discoverOAuthMetadata, getString, sanitizeOAuthMetadata, type OAuthServerMetadata } from '@/lib/connectors/oauth-metadata'
 import { getStrategy } from '@/lib/connectors/oauth-provider-strategies'
 import { OAUTH_CONNECTOR_TYPES, type ConnectorType, type OAuthConnectorType } from '@/lib/connectors/types'
-import { validateConnectorTestEndpoint } from '@/lib/security/ssrf'
 
 type OAuthStatePayload = {
   connectorId: string
@@ -21,13 +21,6 @@ type OAuthStatePayload = {
   registrationEndpoint?: string
   issuer?: string
   mcpServerUrl?: string
-}
-
-type OAuthServerMetadata = {
-  issuer?: string
-  authorizationEndpoint: string
-  tokenEndpoint: string
-  registrationEndpoint?: string
 }
 
 type OAuthClientRegistration = {
@@ -82,29 +75,6 @@ function getOAuthAuthorizeUrlMaxLength(): number {
   const raw = process.env.ARCHE_CONNECTOR_OAUTH_MAX_AUTHORIZE_URL_LENGTH
   const parsed = raw ? Number(raw) : NaN
   return Number.isFinite(parsed) && parsed > 0 ? Math.floor(parsed) : 1800
-}
-
-async function validateConnectorUrl(rawUrl: string): Promise<string> {
-  const validation = await validateConnectorTestEndpoint(rawUrl)
-  if (!validation.ok) {
-    throw new Error(validation.error)
-  }
-  return validation.url.toString()
-}
-
-function getString(value: unknown): string | undefined {
-  return typeof value === 'string' && value.trim() ? value.trim() : undefined
-}
-
-export async function sanitizeOAuthMetadata(metadata: OAuthServerMetadata): Promise<OAuthServerMetadata> {
-  return {
-    issuer: metadata.issuer,
-    authorizationEndpoint: await validateConnectorUrl(metadata.authorizationEndpoint),
-    tokenEndpoint: await validateConnectorUrl(metadata.tokenEndpoint),
-    registrationEndpoint: metadata.registrationEndpoint
-      ? await validateConnectorUrl(metadata.registrationEndpoint)
-      : undefined,
-  }
 }
 
 function resolveOAuthMetadata(
@@ -224,46 +194,6 @@ async function postForm(
   }
 
   return data
-}
-
-export async function discoverOAuthMetadata(mcpServerUrl: string): Promise<OAuthServerMetadata> {
-  const serverUrl = new URL(mcpServerUrl)
-  const authorizationBase = `${serverUrl.protocol}//${serverUrl.host}`
-  const metadataUrl = `${authorizationBase}/.well-known/oauth-authorization-server`
-
-  const metadataResponse = await fetch(metadataUrl, {
-    method: 'GET',
-    headers: {
-      Accept: 'application/json',
-    },
-    cache: 'no-store',
-  }).catch(() => null)
-
-  if (metadataResponse && metadataResponse.ok) {
-    const data = (await metadataResponse.json().catch(() => null)) as Record<string, unknown> | null
-    const authorizationEndpoint = getString(data?.authorization_endpoint)
-    const tokenEndpoint = getString(data?.token_endpoint)
-    if (!authorizationEndpoint || !tokenEndpoint) {
-      throw new Error('oauth_discovery_failed:invalid_metadata')
-    }
-
-    return {
-      issuer: getString(data?.issuer),
-      authorizationEndpoint,
-      tokenEndpoint,
-      registrationEndpoint: getString(data?.registration_endpoint),
-    }
-  }
-
-  if (metadataResponse && metadataResponse.status !== 404) {
-    throw new Error(`oauth_discovery_failed:${metadataResponse.status}`)
-  }
-
-  return {
-    authorizationEndpoint: `${authorizationBase}/authorize`,
-    tokenEndpoint: `${authorizationBase}/token`,
-    registrationEndpoint: `${authorizationBase}/register`,
-  }
 }
 
 async function resolveOAuthPreparationContext(input: {
